// src/types/Gender.ts
export enum Gender {
  Girl = "Girl",
  Boy = "Boy",
  Tomboy = "Tomboy",
  Female = "Female",
  Male = "Male",
  Youngwoman = "Young-woman",
  Youngman = "Young-man",
  Transgirl = "Trans-girl",
  Transboy = "Trans boy",
  Genderfluid = "Gender fluid",
  Agender = "Agender",
  Androgynous = "Androgynous",
  Bigender = "Bi-gender",
  Nonbinary = "Non binary",
  Demiboy = "Demi-boy",
  Demigirl = "Demi-girl",
  Genderqueer = "Genderqueer",
  Gendernonconforming = "Gender non conforming",
  Trigender = "Tri-gender",
  Allgenders = "All genders",
  Inthemiddleofboyandgirl = "In the middle of boy and girl",
  Intersex = "Intersex",
  Notsure = "Not sure",
  Rathernotsay = "Rather not say",
  Others="Others"
}